+++ demo/latest/src/App.tsx
@@ -3,11 +3,13 @@ import Viewer, { Worker } from '@phuocng/react-pdf-viewer';
 
 import '@phuocng/react-pdf-viewer/cjs/react-pdf-viewer.css';
 
+import CanvasLayerExample from './CanvasLayerExample';
+
 const App = () => {
     return (
         <Worker workerUrl="https://unpkg.com/pdfjs-dist@2.2.228/build/pdf.worker.min.js">
             <div style={{ height: '750px', padding: '16px 0' }}>
-                <Viewer fileUrl="/pdf-open-parameters.pdf" />
+                <CanvasLayerExample fileUrl="/pdf-open-parameters.pdf" />
             </div>
         </Worker>
     );